requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==2.0.0
openai==1.35.9
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.0
pydantic==2.8.0
pydantic_core==2.20.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
wheel==0.43.0
anthropic==0.36.2
p-tqdm==1.4.2